Move serial driver to drivers folder.
